requirements.txt
Django
django-admin-interface
django-colorfield
openpyxl
pandas
pillow
PyJWT